.github: Fix logic to skip devcontainer build
In github actions apparently "" == false == 0 and when
we run a workflow with dispatch or "on push" and an inputs is
not defined, the value is "".

So it means in the end that you cannot define a variable that
would default to true because when it's called with dispatch or
"on push" it will be "" and you will not be able to make
the difference between a "" and false.